feat: added feature flag (M2-7240, M2-7243)
