+++ src/modules/Builder/features/ActivityItems/ItemConfiguration/OptionalItemsAndSettings/OptionalItemsAndSettings.const.ts
@@ -1,9 +1,9 @@
 import { ItemResponseType } from 'shared/consts';
 
 import {
-  DEFAULT_MAX_CHARACTERS,
   DEFAULT_DISABLED_TIMER_VALUE,
-  DEFAULT_MAX_CHARACTERS_PARAGRAPH,
+  DEFAULT_MAX_CHARACTERS_SHORT_TEXT,
+  DEFAULT_MAX_CHARACTERS_PARAGRAPH_TEXT,
 } from '../ItemConfiguration.const';
 
 export const DEFAULT_OPTION_VALUE = 0;
@@ -61,7 +61,7 @@ export const defaultSingleSelectionConfig = {
 export const defaultTextConfig = {
   removeBackButton: false,
   skippableItem: false,
-  maxResponseLength: DEFAULT_MAX_CHARACTERS,
+  maxResponseLength: DEFAULT_MAX_CHARACTERS_SHORT_TEXT,
   correctAnswerRequired: false,
   correctAnswer: '',
   numericalResponseRequired: false,
@@ -72,7 +72,7 @@ export const defaultTextConfig = {
 export const defaultParagraphTextConfig = {
   removeBackButton: false,
   skippableItem: false,
-  maxResponseLength: DEFAULT_MAX_CHARACTERS_PARAGRAPH,
+  maxResponseLength: DEFAULT_MAX_CHARACTERS_PARAGRAPH_TEXT,
   responseRequired: false,
 };
