+++ src/modules/Library/features/AppletsCatalog/AppletsCatalog.conts.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_PAGE = 1;
+export const DEFAULT_APPLETS_PER_PAGE = 6;
