+++ src/resources/app-en.json
@@ -1059,7 +1059,8 @@
   "responseOptions_multipleSelection": "Multiple Selection Response Options",
   "responseOptions_singleSelection": "Single Selection Response Option",
   "responseOptions_slider": "Slider Response Options",
-  "responseOptions_text": "Short Text Response Options",
+  "responseOptions_text": "Text Response Options",
+  "responseOptions_shortText": "Short Text Response Options",
   "responseOptions_paragraphText": "Paragraph Text Response Options",
   "responseOptionsTooltip": "The respondent completed the Activity at these times",
   "retainData": "Retain Data",
@@ -1313,8 +1314,10 @@
   "textInputOptionLabel": "Additional Text Input Option",
   "textInputOptionLabel_required": "Additional Text Input Option (Required)",
   "textPlaceholder": "Type text",
-  "textResponseDescription": "The respondent will be able to enter a short text response",
+  "textResponseDescription": "The respondent will be able to enter a text response",
+  "shortTextResponseDescription": "The respondent will be able to enter a short text response",
   "paragraphTextResponseDescription": "The respondent will be able to enter a long text response",
+  "textResponse": "Text Response",
   "shortTextResponse": "Short Text Response",
   "paragraphTextResponse": "Paragraph Text Response",
   "thisReviewerCanView": "This Reviewer Can View",
