+++ src/resources/app-fr.json
@@ -1058,7 +1058,8 @@
   "responseOptions_multipleSelection": "Options de réponse à sélection multiple",
   "responseOptions_singleSelection": "Option de réponse à sélection unique",
   "responseOptions_slider": "Options de réponse avec curseur",
-  "responseOptions_text": "Options de réponse en texte court",
+  "responseOptions_text": "Options de réponse texte",
+  "responseOptions_shortText": "Options de réponse en texte court",
   "responseOptions_paragraphText": "Options de réponse en texte de paragraphe",
   "responseOptionsTooltip": "Le participant a terminé l'activité à ces heures",
   "retainData": "Conserver les données",
@@ -1312,8 +1313,10 @@
   "textInputOptionLabel": "Option de saisie de texte supplémentaire",
   "textInputOptionLabel_required": "Option de saisie de texte supplémentaire (Requis)",
   "textPlaceholder": "Tapez du texte",
-  "textResponseDescription": "Le répondant pourra entrer une courte réponse textuelle",
+  "textResponseDescription": "Le participant pourra saisir une réponse textuelle.",
+  "shortTextResponseDescription": "Le répondant pourra entrer une courte réponse textuelle",
   "paragraphTextResponseDescription": "Le répondant pourra entrer une longue réponse textuelle",
+  "textResponse": "Réponse textuelle",
   "shortTextResponse": "Réponse courte en texte",
   "paragraphTextResponse": "Réponse textuelle longue",
   "thisReviewerCanView": "Ce réviseur peut afficher",
