+++ src/shared/types/featureFlags.ts
@@ -15,6 +15,7 @@ export const FeatureFlagsKeys = {
   enableLorisIntegration: 'enableLorisIntegration',
   enableItemFlowExtendedItems: 'enableItemFlowExtendedItems',
   enableAdminAppSoftLock: 'enableAdminAppSoftLock',
+  enableParagraphTextItem: 'enableParagraphTextItem',
 };
 
 export type FeatureFlags = Partial<Record<keyof typeof FeatureFlagsKeys, LDFlagValue>>;
